+++ services/gradeService.js
@@ -1,4 +1,6 @@
 const gradeRepository = require("../repository/gradeRepository");
+const secondaryResultRepository = require("../repository/secondaryResultRepository");
+const resultRepository = require("../repository/resultRepository");
 
 const getGradeAsync = async() => {
     return await gradeRepository.getGradesAsync();
@@ -13,6 +15,14 @@ const updateGradeAsync = async (grade) => {
 }
 
 const deleteGradeAsync = async (id) => {
+    if (await secondaryResultRepository.hasGradeAsync(id)) {
+        return 0;
+    }
+
+    if (await resultRepository.hasGradeAsync(id)) {
+        return 0;
+    }
+    
     return await gradeRepository.deleteGradeAsync(id);
 }
